Fix TokenPicker
