+++ bridge_ui/src/components/TokenSelectors/EvmTokenPicker.tsx
@@ -20,6 +20,7 @@ import {
   isValidEthereumAddress,
 } from "../../utils/evm";
 import TokenPicker, { BasicAccountRender } from "./TokenPicker";
+import { getTokenLogoAndSymbol } from "../../utils/tokens";
 
 const isWormholev1 = (provider: any, address: string, chainId: ChainId) => {
   if (chainId !== CHAIN_ID_ETH) {
@@ -88,11 +89,17 @@ export default function EvmTokenPicker(
               signerAddress
             );
           } else {
-            return ethTokenToParsedTokenAccount(
+            const logoAndSymbol = await getTokenLogoAndSymbol(chainId, tokenAccount.address)
+            const tokenInfo = await ethTokenToParsedTokenAccount(
               chainId,
               tokenAccount as ethers_contracts.TokenImplementation,
               signerAddress
             );
+            return {
+              ...tokenInfo,
+              symbol: logoAndSymbol?.symbol ?? tokenInfo.symbol,
+              logo: logoAndSymbol?.logoURI ?? tokenInfo.logo
+            }
           }
         } catch (e) {
           return Promise.reject(t("Unable to retrive the specific token."));
