+++ bridge_ui/src/utils/tokens.ts
@@ -1,9 +1,11 @@
 import bscIcon from "../icons/bsc.svg";
 import ethIcon from "../icons/eth.svg";
 import alephiumIcon from "../icons/alephium.svg";
-import { ChainId } from "@alephium/wormhole-sdk";
-import { EXPLORER_API_SERVER_HOST } from "./consts";
+import { CHAIN_ID_ALEPHIUM, CHAIN_ID_BSC, CHAIN_ID_ETH, ChainId, getTokenPoolId, tryNativeToHexString } from "@alephium/wormhole-sdk";
+import { ALEPHIUM_BRIDGE_GROUP_INDEX, ALEPHIUM_TOKEN_BRIDGE_CONTRACT_ID, EXPLORER_API_SERVER_HOST } from "./consts";
 import i18n from "../i18n";
+import { getAlephiumTokenLogoAndSymbol } from "./alephium";
+import { getBSCTokenLogoAndSymbol, getETHTokenLogoAndSymbol } from "./evm";
 
 export type RegisteredTokenInfo = {
   tokenAddress: string
@@ -40,5 +42,28 @@ export async function getRegisteredTokens(): Promise<RegisteredTokenInfo[]> {
   } catch (error) {
     console.log(error)
     return []
+  }
+}
+
+export async function getTokenLogoAndSymbol(tokenChainId: ChainId, tokenId: string): Promise<{ logoURI?: string, symbol?: string } | undefined> {
+  if (tokenChainId !== CHAIN_ID_ALEPHIUM) {
+    const wrappedIdOnALPH = getTokenPoolId(
+      ALEPHIUM_TOKEN_BRIDGE_CONTRACT_ID,
+      tokenChainId,
+      tryNativeToHexString(tokenId, tokenChainId),
+      ALEPHIUM_BRIDGE_GROUP_INDEX
+    )
+    const info = await getAlephiumTokenLogoAndSymbol(wrappedIdOnALPH)
+    if (info !== undefined) return info
+  }
+
+  if (tokenChainId === CHAIN_ID_ETH) {
+    return getETHTokenLogoAndSymbol(tokenId)
+  } else if (tokenChainId === CHAIN_ID_ALEPHIUM) {
+    return getAlephiumTokenLogoAndSymbol(tokenId)
+  } else if (tokenChainId === CHAIN_ID_BSC) {
+    return getBSCTokenLogoAndSymbol(tokenId)
+  } else {
+    return undefined
   }
 }
